Fix imports, rename behavior
