+++ packages/fluentui/docs/src/examples/components/Card/Usage/CardExampleFocusableChildren.tsx
@@ -1,7 +1,7 @@
 import {
   Button,
   Card,
-  cardMultipleFocusableBehavior,
+  cardChildrenFocusableBehavior,
   DownloadIcon,
   Flex,
   MoreIcon,
@@ -12,7 +12,7 @@ import * as React from 'react';
 
 const CardExampleFocusableChildren = () => {
   return (
-    <Card accessibility={cardMultipleFocusableBehavior} aria-roledescription="card with action buttons">
+    <Card accessibility={cardChildrenFocusableBehavior} aria-roledescription="card with action buttons">
       <Card.Header>
         <Text content="Sample title" weight="bold" />
       </Card.Header>
